+++ src/GeniusYield/Transaction/CoinSelection/Balance.hs
@@ -19,6 +19,7 @@ import Cardano.Api qualified as Api
 import Cardano.Api.Shelley qualified as Api.S
 import Cardano.Ledger.Binary qualified as CBOR
 import Cardano.Ledger.Conway.Core (eraProtVerHigh)
+import Cardano.Numeric.Util (padCoalesce, partitionNatural)
 import Control.Monad.Extra (
   andM,
   (<=<),
@@ -31,6 +32,9 @@ import Control.Monad.Trans.Except (
 import Data.Bifunctor qualified
 import Data.ByteString qualified as BS
 import Data.Default (Default (def))
+import Data.Either.Extra (
+  maybeToEither,
+ )
 import Data.Foldable qualified as F
 import Data.Generics.Internal.VL.Lens (
   view,
@@ -41,6 +45,8 @@ import Data.List.NonEmpty (
  )
 import Data.List.NonEmpty qualified as NE
 import Data.Map qualified as Map
+import Data.Monoid (Sum (..))
+import Data.Ord (comparing)
 import Data.Semigroup (mtimesDefault)
 import Data.Set qualified as S
 import Data.Set qualified as Set
@@ -69,16 +75,18 @@ data ValueSizeAssessment
     ValueSizeExceedsLimit
   deriving (Eq, Generic, Show)
 
+type ValueSizeAssessor = GYValue -> ValueSizeAssessment
+
 -- TODO: Having this 'v' lingering around is nuisance.
 
 -- | Specifies all constraints required for coin selection.
 data SelectionConstraints v = SelectionConstraints
   { tokenBundleSizeAssessor ::
-      GYValue ->
-      ValueSizeAssessment
+      ValueSizeAssessor
   -- ^ Assesses the size of a value relative to the upper limit of
   -- what can be included in a transaction output.
   -- TODO: Make it just a Bool?
+  -- TODO: Get rid of "bundle"...
   , computeMinimumAdaQuantity ::
       GYAddress ->
       GYValue ->
@@ -608,7 +616,7 @@ performSelectionEmpty performSelectionFn constraints params =
   transformResult r@SelectionResult {..} =
     r
       { extraCoinSource =
-          transform (\ecs -> if ecs >= dummyCoin then ecs - dummyCoin else 0) (const id) extraCoinSource
+          transform (`naturalMonus` dummyCoin) (const id) extraCoinSource
       , outputsCovered =
           transform (const []) (const . F.toList) outputsCovered
       }
@@ -1154,7 +1162,807 @@ data MakeChangeCriteria minCoinFor bundleSizeAssessor = MakeChangeCriteria
   }
   deriving (Eq, Generic, Show)
 
-makeChange = undefined
+{- | Indicates 'True' if and only if a token bundle exceeds the maximum size
+  that can be included in a transaction output.
+-}
+tokenBundleSizeExceedsLimit :: ValueSizeAssessor -> GYValue -> Bool
+tokenBundleSizeExceedsLimit assess v =
+  case assess v of
+    ValueSizeWithinLimit ->
+      False
+    ValueSizeExceedsLimit ->
+      True
+
+{- | Constructs change bundles for a set of selected inputs and outputs.
+
+Returns 'Nothing' if the specified inputs do not provide enough ada to
+satisfy the minimum delta and minimum ada quantities of the change bundles
+generated.
+
+This function will generate runtime errors if:
+
+   1.  The total balance of all outputs is not less than or equal to the
+       total balance of all inputs.
+
+   2.  The total ada balance of all outputs is zero.
+
+Pre-condition (1) should be satisfied by any result produced by the
+'runSelection' function.
+
+Pre-condition (2) should be satisfied by assigning a minimum ada quantity
+to every output token bundle.
+-}
+makeChange ::
+  -- | Criteria for making change.
+  MakeChangeCriteria (GYValue -> Natural) ValueSizeAssessor ->
+  -- | Generated change bundles.
+  Either UnableToConstructChangeError [GYValue]
+makeChange criteria
+  | not (totalOutputValue `valueLessOrEqual` totalInputValue) =
+      totalInputValueInsufficient
+  | valueAda totalOutputValue == 0 =
+      totalOutputCoinValueIsZero
+  | otherwise =
+      first mkUnableToConstructChangeError $ do
+        adaAvailable <-
+          maybeToEither
+            (requiredCost `naturalMonus` excessCoin)
+            (excessCoin `naturalReductive` requiredCost)
+        assignCoinsToChangeMaps
+          adaAvailable
+          minCoinFor
+          changeMapOutputCoinPairs
+ where
+  MakeChangeCriteria
+    { minCoinFor
+    , bundleSizeAssessor
+    , requiredCost
+    , extraCoinSource
+    , extraCoinSink
+    , inputBundles
+    , outputBundles
+    , assetsToMint
+    , assetsToBurn
+    , maximumOutputAdaQuantity
+    , maximumOutputTokenQuantity
+    } = criteria
+
+  -- The following subtraction is safe, as we have already checked
+  -- that the total input value is greater than the total output
+  -- value:
+  excess :: GYValue
+  excess = totalInputValue `valueMonus` totalOutputValue
+
+  (fromInteger -> excessCoin, valueToList -> excessAssets) = valueSplitAda excess
+
+  -- Change maps for all assets, where each change map is paired with a
+  -- corresponding coin from the original outputs.
+  --
+  -- When combining change maps from user-specified assets and non-user-
+  -- specified assets, we arrange that any empty maps in either list are
+  -- combined together if possible, so as to give 'assignCoinsToChangeMaps'
+  -- the greatest chance of success.
+  --
+  -- This list is sorted into ascending order of asset count, where empty
+  -- change maps are all located at the start of the list.
+  --
+  changeMapOutputCoinPairs :: NonEmpty (GYValue, Natural) -- First component is actually tokens without ada.
+  changeMapOutputCoinPairs =
+    outputCoins
+      -- First, combine the original output coins with the change maps for
+      -- user-specified assets. We must pair these together right at the
+      -- start in order to retain proportionality with the original outputs.
+      & NE.zip changeForUserSpecifiedAssets
+      -- Next, sort the list into ascending order of asset count, which moves
+      -- any empty maps to the start of the list:
+      & NE.sortWith (AssetCount . fst)
+      -- Next, combine the existing list with the change maps for non-user
+      -- specified assets, which are already sorted into ascending order of
+      -- asset count:
+      & NE.zipWith
+        (\m1 (m2, c) -> (m1 <> m2, c))
+        changeForNonUserSpecifiedAssets
+      -- Finally, if there are any maps that are oversized (in any way), then
+      -- split these maps up along with their corresponding output coins:
+      & splitOversizedMaps
+   where
+    splitOversizedMaps ::
+      NonEmpty (GYValue, Natural) -> NonEmpty (GYValue, Natural)
+    splitOversizedMaps =
+      -- For the sake of convenience when splitting up change maps and
+      -- output coins (which are treated as weights), treat each change
+      -- map and its corresponding output coin as a token bundle.
+      fmap unbundle . split . fmap bundle
+     where
+      bundle (m, c) = m <> valueFromLovelace (fromIntegral c)
+      unbundle val = let (c, m) = valueSplitAda val in (m, fromIntegral c)
+      split b =
+        b
+          & flip
+            splitBundlesWithExcessiveAssetCounts
+            (tokenBundleSizeExceedsLimit assessBundleSizeWithMaxCoin)
+          & flip
+            splitBundlesWithExcessiveTokenQuantities
+            maximumOutputTokenQuantity
+
+      -- When assessing the size of a change map to determine if it is
+      -- excessively large, we don't yet know how large the associated
+      -- ada quantity will be, since ada quantities are assigned at a
+      -- later stage (in 'assignCoinsToChangeMaps').
+      --
+      -- Therefore, we err on the side of caution, and assess the size
+      -- of a change map combined with the maximum possible ada quantity.
+      --
+      -- This means that when presented with a very large change map, we
+      -- have a small chance of splitting the map even if that map would
+      -- be within the limit when combined with its final ada quantity.
+      --
+      -- However, oversplitting a change map is preferable to creating
+      -- a bundle that is marginally over the limit, which would cause
+      -- the resultant transaction to be rejected.
+      --
+      assessBundleSizeWithMaxCoin :: ValueSizeAssessor
+      assessBundleSizeWithMaxCoin =
+        bundleSizeAssessor
+          . valueInsert GYLovelace (fromIntegral maximumOutputAdaQuantity)
+
+  -- Change for user-specified assets: assets that were present in the
+  -- original set of user-specified outputs ('outputsToCover').
+  changeForUserSpecifiedAssets :: NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+  changeForUserSpecifiedAssets =
+    F.foldr
+      ( NE.zipWith (<>)
+          . makeChangeForUserSpecifiedAsset outputMaps
+      )
+      (mempty <$ outputMaps)
+      (fmap (Data.Bifunctor.second fromInteger) excessAssets)
+
+  -- Change for non-user-specified assets: assets that were not present
+  -- in the original set of user-specified outputs ('outputsToCover').
+  changeForNonUserSpecifiedAssets :: NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+  changeForNonUserSpecifiedAssets =
+    makeChangeForNonUserSpecifiedAssets
+      outputMaps
+      nonUserSpecifiedAssetQuantities
+      & addMintValuesToChangeMaps
+        (removeUserSpecifiedAssetIds assetsToMint)
+      & removeBurnValuesFromChangeMaps
+        (removeUserSpecifiedAssetIds assetsToBurn)
+   where
+    removeUserSpecifiedAssetIds :: GYValue -> GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+    removeUserSpecifiedAssetIds = flip valueWithoutAssets userSpecifiedAssetIds
+
+  totalInputValueInsufficient =
+    error
+      "makeChange: not (totalOutputValue <= totalInputValue)"
+  totalOutputCoinValueIsZero =
+    error
+      "makeChange: not (totalOutputCoinValue > 0)"
+
+  mkUnableToConstructChangeError :: Natural -> UnableToConstructChangeError
+  mkUnableToConstructChangeError shortfall =
+    UnableToConstructChangeError
+      { requiredCost
+      , shortfall
+      }
+
+  outputMaps :: NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+  outputMaps = valueNonAda <$> outputBundles
+
+  outputCoins :: NonEmpty Natural
+  outputCoins = fromInteger . valueAda <$> outputBundles
+
+  totalInputValue :: GYValue
+  totalInputValue =
+    F.fold inputBundles
+      <> valueFromLovelace (fromIntegral extraCoinSource)
+      -- Mints represent extra inputs from "the void"
+      <> assetsToMint
+
+  totalOutputValue :: GYValue
+  totalOutputValue =
+    F.fold outputBundles
+      <> valueFromLovelace (fromIntegral extraCoinSink)
+      -- Burns represent extra outputs to "the void"
+      <> assetsToBurn
+
+  -- Identifiers of all user-specified assets: assets that were included in
+  -- the original set of outputs.
+  userSpecifiedAssetIds :: Set GYAssetClass
+  userSpecifiedAssetIds = valueAssets (F.fold outputBundles)
+
+  -- Identifiers and quantities of all non-user-specified assets: assets that
+  -- were not included in the original set of outputs, but that were
+  -- nevertheless selected during the selection process.
+  --
+  -- Each asset is paired with the complete list of quantities of that asset
+  -- present in the selected inputs.
+  nonUserSpecifiedAssetQuantities :: Map GYAssetClass (NonEmpty Natural)
+  nonUserSpecifiedAssetQuantities =
+    collateNonUserSpecifiedAssetQuantities
+      (valueNonAda <$> inputBundles)
+      userSpecifiedAssetIds
+
+{- | Generates a map of all non-user-specified assets and their quantities.
+
+Each key in the resulting map corresponds to an asset that was NOT included
+in the original set of user-specified outputs, but that was nevertheless
+selected during the selection process.
+
+The value associated with each key corresponds to the complete list of all
+discrete non-zero quantities of that asset present in the selected inputs.
+-}
+collateNonUserSpecifiedAssetQuantities ::
+  -- | Token maps of all selected inputs.
+  NonEmpty GYValue -> -- Denotes without lovelace.
+
+  -- | Set of all assets in user-specified outputs.
+  Set GYAssetClass ->
+  Map GYAssetClass (NonEmpty Natural)
+collateNonUserSpecifiedAssetQuantities inputMaps userSpecifiedAssetIds =
+  F.foldr discardUserSpecifiedAssets mempty inputMaps
+ where
+  discardUserSpecifiedAssets ::
+    GYValue -> -- Denotes without lovelace.
+    Map GYAssetClass (NonEmpty Natural) ->
+    Map GYAssetClass (NonEmpty Natural)
+  discardUserSpecifiedAssets tokens m =
+    foldr (\(k, fromIntegral -> v) -> Map.insertWith (<>) k (v :| [])) m filtered
+   where
+    filtered =
+      filter
+        ((`Set.notMember` userSpecifiedAssetIds) . fst)
+        (valueToList tokens)
+
+{- | Assigns coin quantities to a list of pre-computed asset change maps.
+
+Each pre-computed asset change map must be paired with the original coin
+value of its corresponding output.
+
+This function:
+
+   - expects the list of pre-computed asset change maps to be sorted in an
+     order that ensures all empty token maps are at the start of the list.
+
+   - attempts to assign a minimum ada quantity to every change map, but
+     iteratively drops empty change maps from the start of the list if the
+     amount of ada is insufficient to cover them all.
+
+   - continues dropping empty change maps from the start of the list until
+     it is possible to assign a minimum ada value to all remaining entries.
+
+   - returns a list that is identical in length to the input list if (and
+     only if) it was possible to assign a minimum ada quantity to all change
+     maps.
+
+   - returns a list that is shorter than the input list if it was only
+     possible to assign a minimum ada quantity to a suffix of the given
+     list.
+
+   - fails if (and only if) there was not enough ada available to assign the
+     minimum ada quantity to all non-empty change maps.
+-}
+assignCoinsToChangeMaps ::
+  HasCallStack =>
+  -- | The total quantity of ada available, including any extra source of ada.
+  Natural ->
+  -- | A function to calculate the minimum required ada quantity for any
+  -- token map.
+  (GYValue -> Natural) -> -- Denotes without lovelace
+
+  -- | A list of pre-computed asset change maps paired with original output
+  -- coins, sorted into an order that ensures all empty token maps are at the
+  -- start of the list.
+  NonEmpty (GYValue, Natural) -> -- Denotes without lovelace.
+
+  -- | Resulting change bundles, or the shortfall quantity if there was not
+  -- enough ada available to assign a minimum ada quantity to all non-empty
+  -- token maps.
+  Either Natural [GYValue]
+assignCoinsToChangeMaps adaAvailable minCoinFor pairsAtStart
+  | not changeMapsCorrectlyOrdered =
+      changeMapsNotCorrectlyOrderedError
+  | otherwise =
+      loop adaRequiredAtStart pairsAtStart
+ where
+  loop !adaRequired !pairsNonEmpty = case pairsNonEmpty of
+    pair :| pairs
+      | adaAvailable >= adaRequired ->
+          -- We have enough ada available to pay for the minimum required
+          -- amount of every asset map that remains in our list:
+          let
+            assetMapsRemaining = fst <$> (pair :| pairs)
+            bundlesForAssetsWithMinimumCoins =
+              assignMinimumCoin minCoinFor <$> assetMapsRemaining
+            -- Calculate the amount of ada that remains after assigning the
+            -- minimum amount to each map. This should be safe, as we have
+            -- already determined that we have enough ada available:
+            adaRemaining = adaAvailable `naturalMonus` adaRequired
+            -- Partition any remaining ada according to the weighted
+            -- distribution of output coins that remain in our list:
+            outputCoinsRemaining = snd <$> (pair :| pairs)
+            bundlesForOutputCoins =
+              valueFromLovelace . fromIntegral
+                <$> makeChangeForCoin outputCoinsRemaining adaRemaining
+           in
+            -- Finally, combine the minimal coin asset bundles with the
+            -- bundles obtained by partitioning the remaining ada amount:
+            Right $
+              NE.toList $
+                NE.zipWith
+                  (<>)
+                  bundlesForAssetsWithMinimumCoins
+                  bundlesForOutputCoins
+    (m, _) :| (p : ps)
+      | isEmptyValue m && adaAvailable < adaRequired ->
+          -- We don't have enough ada available to pay for the minimum
+          -- required amount of every asset map, but we do have an empty
+          -- asset map that is safe to drop. This will reduce the amount of
+          -- ada required by a small amount:
+          let adaRequired' = adaRequired `naturalMonus` minCoinFor m
+           in loop adaRequired' (p :| ps)
+    (m, _) :| []
+      | isEmptyValue m && adaAvailable < adaRequired ->
+          -- We didn't have any non-ada assets at all in our change, and we
+          -- also don't have enough ada available to pay even for a single
+          -- change output. We just burn the available ada amount (which
+          -- will be small), returning no change.
+          Right []
+    _ ->
+      -- We don't have enough ada available, and there are no empty token
+      -- maps available to drop. We have to give up at this point.
+      Left (adaRequired `naturalMonus` adaAvailable)
+
+  adaRequiredAtStart = getSum $ F.fold $ Sum . minCoinFor . fst <$> pairsAtStart
+
+  changeMaps = fst <$> pairsAtStart
+
+  -- Indicates whether or not the given change maps are correctly ordered,
+  -- so that all empty maps are located at the start of the list.
+  changeMapsCorrectlyOrdered =
+    (==)
+      (NE.takeWhile isEmptyValue changeMaps)
+      (NE.filter isEmptyValue changeMaps)
+
+  changeMapsNotCorrectlyOrderedError =
+    error $
+      unwords
+        [ "assignCoinsToChangeMaps: pre-computed asset change maps must be"
+        , "arranged in an order where all empty maps are at the start of"
+        , "the list."
+        ]
+
+-- | Assigns the minimum required ada quantity to a token map.
+assignMinimumCoin :: (GYValue -> Natural) -> GYValue -> GYValue -- This first two 'GYValue' in this signature denote tokens without lovelace.
+assignMinimumCoin minCoinFor m = valueFromLovelace (fromIntegral $ minCoinFor m) <> m
+
+{- | Constructs change outputs for a user-specified asset: an asset that was
+  present in the original set of outputs.
+
+If the given asset does not appear in the given distribution, this function
+returns a list of empty token maps. Otherwise, the given token quantity is
+partitioned into a list of quantities that are proportional to the weights
+within the given input distribution, modulo rounding.
+
+The length of the output list is always the same as the the length of the
+input list, and the sum of its quantities is either zero, or exactly equal
+to the token quantity in the second argument.
+-}
+makeChangeForUserSpecifiedAsset ::
+  -- | A list of weights for the distribution. Conveniently captures both
+  -- the weights, and the number of elements amongst which the quantity
+  -- should be distributed.
+  NonEmpty GYValue -> -- Denotes without lovelace.
+
+  -- | A surplus token quantity to distribute.
+  (GYAssetClass, Natural) ->
+  NonEmpty GYValue
+makeChangeForUserSpecifiedAsset targets (asset, excess) =
+  valueSingleton asset
+    <$> maybe (fmap fromIntegral zeros) (fmap fromIntegral) (partitionNatural excess weights)
+ where
+  weights :: NonEmpty Natural
+  weights = fromIntegral . flip valueAssetClass asset <$> targets
+
+  zeros :: NonEmpty Natural
+  zeros = 0 <$ targets
+
+{- | Constructs change outputs for a non-user-specified asset: an asset that
+  was not present in the original set of outputs.
+
+This function constructs a list of change outputs by preserving the input
+distribution as much as possible. Note that only the length of the first
+argument is used.
+
+The length of the output list is always the same as the length of the input
+list, and the sum of its quantities is always exactly equal to the sum of
+all token quantities given in the second argument.
+
+The resultant list is sorted into ascending order when maps are compared
+with the `leq` function.
+-}
+makeChangeForNonUserSpecifiedAsset ::
+  -- | Determines the number of change maps to create.
+  NonEmpty a ->
+  -- | An asset quantity to distribute.
+  (GYAssetClass, NonEmpty Natural) ->
+  -- | The resultant change maps.
+  NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+makeChangeForNonUserSpecifiedAsset n (asset, fmap (Sum . fromIntegral) -> quantities) =
+  valueSingleton asset . getSum <$> padCoalesce quantities n
+
+{- | Constructs change outputs for all non-user-specified assets: assets that
+  were not present in the original set of outputs.
+
+The resultant list is sorted into ascending order when maps are compared
+with the `leq` function.
+-}
+makeChangeForNonUserSpecifiedAssets ::
+  -- | Determines the number of change maps to create.
+  NonEmpty a ->
+  -- | A map of asset quantities to distribute.
+  Map GYAssetClass (NonEmpty Natural) ->
+  -- | The resultant change maps.
+  NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+makeChangeForNonUserSpecifiedAssets n nonUserSpecifiedAssetQuantities =
+  F.foldr
+    (NE.zipWith (<>) . makeChangeForNonUserSpecifiedAsset n)
+    (mempty <$ n)
+    (Map.toList nonUserSpecifiedAssetQuantities)
+
+{- | Constructs a list of ada change outputs based on the given distribution.
+
+If the sum of weights in given distribution is equal to zero, this function
+throws a runtime error.
+
+The length of the output list is always the same as the length of the input
+list, and the sum of its quantities is always exactly equal to the 'Coin'
+value given as the second argument.
+-}
+makeChangeForCoin ::
+  HasCallStack =>
+  -- | A list of weights for the distribution. Conveniently captures both
+  -- the weights, and the number of elements amongst which the surplus
+  -- ada quantity should be distributed.
+  NonEmpty Natural ->
+  -- | A surplus ada quantity to be distributed.
+  Natural ->
+  NonEmpty Natural
+makeChangeForCoin ws c = fromMaybe zeroWeightSumError $ partitionNatural c ws
+ where
+  zeroWeightSumError =
+    error
+      "makeChangeForCoin: weights must have a non-zero sum."
+
+--------------------------------------------------------------------------------
+-- Minting and burning
+--------------------------------------------------------------------------------
+
+-- Once we know how much change to give, grouping the change into bundles is a
+-- somewhat complicated topic.
+--
+-- We want to create change outputs with, as far as possible, values that are
+-- likely to be useful to the user in future, where values that more closely
+-- approximate the user-specified outputs are considered more "useful".
+--
+-- A key property is that the number of change outputs should reflect the
+-- number of outputs specified by the user. For example, if the user sends
+-- value to five distinct outputs, we should create five distinct change
+-- outputs.
+--
+-- However, we also want to mint and burn tokens. In general, minting tokens
+-- requires us to add value to the change outputs and burning tokens requires
+-- us to remove value from the change outputs.
+--
+-- It's also important to note that the change bundle calculation requires
+-- that the change for user-specified and non-user-specified assets have the
+-- following properties:
+--
+--    1. The lists share the same length;
+--    2. The lists are in ascending partial order.
+--
+-- For example, given the following non-user-specified asset quantities:
+--
+--    [ ("A", [4, 1, 3, 2])
+--    , ("B", [9, 1, 8, 2, 7, 3, 6, 4, 5])
+--    ]
+--
+-- If the user requests 5 outputs in their transaction,
+-- 'makeChangeForNonUserSpecifiedAssets' will generate:
+--
+--    [ [          ("B",  7) ]
+--    [ [("A", 1), ("B",  8) ]
+--    [ [("A", 2), ("B",  9) ]
+--    [ [("A", 3), ("B",  9) ]
+--    [ [("A", 4), ("B", 12) ]
+--
+-- That is to say, it generates change bundles that satisfy the following
+-- properties:
+--
+--    1.  The number of change bundles matches the number of outputs
+--        the user originally requested;
+--    2.  The change bundles are split in such a way to maximize the
+--        number of large change bundles.
+--
+-- The change function maintains the property that the change bundles are in
+-- ascending partial order, such that each change bundle is a subset of the
+-- next. This property is required by 'changeMapOutputCoinPairs', so it's
+-- important it's maintained.
+--
+-- The following two functions work by modifying the change bundles for
+-- non-user-specified assets.
+--
+-- We add minted tokens to the largest change bundle:
+--
+--    [ [          ("B",  7) ]
+--    [ [("A", 1), ("B",  8) ]
+--    [ [("A", 2), ("B",  9) ]
+--    [ [("A", 3), ("B",  9) ]
+--    [ [("A", 4), ("B", 12) ] <-- add minted tokens here
+--
+-- We remove burned tokens from the smallest change bundles, until all burned
+-- tokens are removed:
+--
+--    [ [          ("B", 7)  ] <-- start removing burned tokens from here
+--    [ [("A", 1), ("B", 8)  ] <-- if we must burn more, remove from here
+--    [ [("A", 2), ("B", 9)  ] <-- if we must burn more, remove from here
+--    [ [("A", 3), ("B", 9)  ] <-- and so on, until we've removed everything.
+--    [ [("A", 4), ("B", 12) ]
+--
+-- The solution for minting maintains the properties we desire, namely:
+--
+--    1.  The number of change bundles matches the number of
+--        "outputs to cover" (we are not changing the number of bundles).
+--    2.  The change bundles are in ascending partial order (by adding to the
+--        largest bundle we trivially maintain ordering).
+--    3.  The change bundles are split in such a way to maximize the
+--        number of large change bundles.
+--
+-- The solution for burning maintains the same properties:
+--
+--    1.  The number of change bundles is not changed, in the case we burn a
+--        change bundle completely, we just leave it as an empty entry
+--        (effectively "pad with zeros").
+--    2.  By removing from the smallest bundle, we maintain the ascending
+--        partial order of the change bundles.
+--    3.  By removing from the smallest bundles, we remove the "least useful"
+--        bundles, maximizing the overall usefulness of our bundles.
+
+{- | Adds a minted asset quantity to a list of change maps.
+
+This function always adds the given quantity to the final change map in the
+given list.
+
+Example:
+
+Suppose we have the following list of change maps:
+
+   [ [          ("B",  7) ]
+   [ [("A", 1), ("B",  8) ]
+   [ [("A", 2), ("B",  9) ]
+   [ [("A", 3), ("B",  9) ]
+   [ [("A", 4), ("B", 12) ]
+
+If we add 4 tokens of asset "A", we obtain the following result:
+
+   [ [          ("B",  7) ]
+   [ [("A", 1), ("B",  8) ]
+   [ [("A", 2), ("B",  9) ]
+   [ [("A", 3), ("B",  9) ]
+   [ [("A", 8), ("B", 12) ] -- Increased by 4
+
+Provided that the specified change maps are in ascending partial order, this
+function guarantees that the resulting change maps will also be in ascending
+partial order.
+
+The length of the given list is preserved in the output list.
+-}
+addMintValueToChangeMaps ::
+  (GYAssetClass, Natural) ->
+  NonEmpty GYValue -> -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+  NonEmpty GYValue -- Actually denotes tokens without ada though this is not captured by our GYValue type.
+addMintValueToChangeMaps (assetId, fromIntegral -> assetQty) =
+  -- The largest element is the last element in an ascending order list
+  modifyLast $ \m -> valueAdjust (assetQty +) assetId m
+ where
+  modifyLast f xs = case NE.reverse xs of
+    (y :| ys) -> NE.reverse (f y :| ys)
+
+{- | Adds minted values for multiple assets to a list of change maps.
+
+Plural of @addMintValueToChangeMaps@.
+-}
+addMintValuesToChangeMaps ::
+  -- | Map of minted values
+  GYValue ->
+  -- | Change maps
+  NonEmpty GYValue ->
+  -- | Change maps with minted values
+  NonEmpty GYValue -- All 'GYValue' here are for without lovelace.
+addMintValuesToChangeMaps =
+  flip (F.foldr addMintValueToChangeMaps) . map (Data.Bifunctor.second fromInteger) . valueToList
+
+{- | Removes a burned asset quantity from a list of change maps.
+
+For a given asset 'a' and reduction target 't', this function traverses the
+given list from left to right, reducing the quantity of asset 'a' in each
+change map until the reduction target 't' has been met, or until the list
+is exhausted.
+
+For each change map 'm' under consideration:
+
+   - if the quantity 'q' of asset 'a' in map 'm' is less than or equal to
+     the remaining required reduction 'r', it will be replaced with a zero
+     (effectively eliminating asset 'a' from the map).
+
+   - if the quantity 'q' of asset 'a' in map 'm' is greater than the
+     remaining required reduction 'r', it will be replaced with the
+     absolute difference between 'q' and 'r'.
+
+If the total quantity of the given asset in the given change maps is greater
+than the specified reduction target, the total reduction will be equal to
+the specified reduction target. Otherwise, the given asset will be
+completely eliminated from all change maps.
+
+Example:
+
+Suppose we have the following list of change maps:
+
+   [ [          ("B",  7) ]
+   [ [("A", 1), ("B",  8) ]
+   [ [("A", 2), ("B",  9) ]
+   [ [("A", 3), ("B",  9) ]
+   [ [("A", 4), ("B", 12) ]
+
+If our target is to reduce the quantity of asset "A" by 4, then we should
+obtain the following result:
+
+   [ [          ("B",  7) ] -- Unable to reduce (already 0)
+   [ [          ("B",  8) ] -- Reduced by 1 (and eliminated from map)
+   [ [          ("B",  9) ] -- Reduced by 2 (and eliminated from map)
+   [ [("A", 2), ("B",  9) ] -- Reduced by 1
+   [ [("A", 4), ("B", 12) ]
+
+Provided that the specified change maps are in ascending partial order, this
+function guarantees that the resulting change maps will also be in ascending
+partial order.
+
+The length of the given list is preserved in the output list.
+-}
+removeBurnValueFromChangeMaps ::
+  -- | Asset quantity reduction target
+  (GYAssetClass, Natural) ->
+  -- | Change maps with quantities of the given asset to be reduced
+  NonEmpty GYValue -> -- without lovelace
+
+  -- | Change maps with reduced quantities of the given asset
+  NonEmpty GYValue -- without lovelace
+removeBurnValueFromChangeMaps (assetId, assetQty) maps =
+  maps
+    & fmap (fromIntegral . (`valueAssetClass` assetId))
+    & reduceTokenQuantities assetQty
+    & NE.zipWith (\val n -> valueInsert assetId (fromIntegral n) val) maps
+
+{- | Reduces the total value of the given list of token quantities by the given
+  reduction target.
+
+This function traverses the given list of quantities from left to right,
+reducing each quantity in turn until the total reduction is equal to the
+given reduction target, or until the list is exhausted.
+
+For each quantity 'q' under consideration:
+
+   - if 'q' is less than or equal to the remaining required reduction 'r',
+     it will be replaced with a zero.
+
+   - if 'q' is greater than the remaining required reduction 'r', it will
+     be replaced with the absolute difference between 'q' and 'r'.
+
+If the total value in the list is less than the reduction target, the
+result will be a list of zeros.
+
+Provided the given list is in ascending order, the resulting list is also
+guaranteed to be in ascending order.
+
+The length of the given list is preserved in the output.
+-}
+reduceTokenQuantities ::
+  -- | Reduction target
+  Natural ->
+  -- | List of quantities to reduce
+  NonEmpty Natural ->
+  -- | The list of reduced quantities
+  NonEmpty Natural
+reduceTokenQuantities reductionTarget quantities =
+  NE.fromList $ burn reductionTarget (NE.toList quantities) []
+ where
+  burn _ [] ys = reverse ys
+  burn b (x : xs) ys
+    | x >= b = reverse ys <> (x' : xs)
+    | otherwise = burn b' xs (x' : ys)
+   where
+    b' = b `naturalMonus` x
+    x' = x `naturalMonus` b
+
+{- | Removes burned values for multiple assets from a list of change maps.
+
+Plural of @removeBurnValueFromChangeMaps@.
+-}
+removeBurnValuesFromChangeMaps ::
+  -- | Map of burned values
+  GYValue -> -- all 'GYValue' here are for without lovelace.
+
+  -- | Change maps
+  NonEmpty GYValue ->
+  -- | Change maps with burned values removed
+  NonEmpty GYValue
+removeBurnValuesFromChangeMaps =
+  flip (F.foldr removeBurnValueFromChangeMaps) . fmap (Data.Bifunctor.second fromIntegral) . valueToList
+
+--------------------------------------------------------------------------------
+-- Splitting bundles
+--------------------------------------------------------------------------------
+
+{- | Splits a bundle into smaller bundles if its asset count is excessive when
+  measured with the given 'isExcessive' indicator function.
+
+Returns a list of smaller bundles for which 'isExcessive' returns 'False'.
+-}
+splitBundleIfAssetCountExcessive ::
+  -- | The token bundle suspected to have an excessive number of assets.
+  GYValue ->
+  -- | A function that returns 'True' if (and only if) the asset count of
+  -- the given bundle is excessive.
+  (GYValue -> Bool) ->
+  NonEmpty GYValue
+splitBundleIfAssetCountExcessive b isExcessive
+  | isExcessive b =
+      splitInHalf b >>= flip splitBundleIfAssetCountExcessive isExcessive
+  | otherwise =
+      pure b
+ where
+  splitInHalf = flip equipartitionAssets (() :| [()])
+
+equipartitionAssets = undefined
+equipartitionQuantitiesWithUpperBound = undefined
+
+{- | Splits bundles with excessive asset counts into smaller bundles.
+
+Only token bundles where the 'isExcessive' indicator function returns 'True'
+will be split.
+
+Returns a list of smaller bundles for which 'isExcessive' returns 'False'.
+
+If none of the bundles in the given list has an excessive asset count,
+this function will return the original list.
+-}
+splitBundlesWithExcessiveAssetCounts ::
+  -- | Token bundles.
+  NonEmpty GYValue ->
+  -- | A function that returns 'True' if (and only if) the asset count of
+  -- the given bundle is excessive.
+  (GYValue -> Bool) ->
+  NonEmpty GYValue
+splitBundlesWithExcessiveAssetCounts bs isExcessive =
+  (`splitBundleIfAssetCountExcessive` isExcessive) =<< bs
+
+{- | Splits bundles with excessive token quantities into smaller bundles.
+
+Only token bundles containing quantities that exceed the maximum token
+quantity will be split.
+
+If none of the bundles in the given list contain a quantity that exceeds
+the maximum token quantity, this function will return the original list.
+-}
+splitBundlesWithExcessiveTokenQuantities ::
+  -- | Token bundles.
+  NonEmpty GYValue ->
+  -- | Maximum allowable token quantity.
+  Natural ->
+  -- | The partitioned bundles.
+  NonEmpty GYValue
+splitBundlesWithExcessiveTokenQuantities bs maxQuantity =
+  (`equipartitionQuantitiesWithUpperBound` maxQuantity) =<< bs
 
 --------------------------------------------------------------------------------
 -- Round-robin processing
@@ -1192,6 +2000,24 @@ selectedCoinQuantity =
     . UTxOSelection.selectedBalance
 
 --------------------------------------------------------------------------------
+-- Utility types
+--------------------------------------------------------------------------------
+
+{- | A total ordering on value based on the number of assets in it.
+
+If two values have the same number of assets, then we fall back to ordinary
+ordering as a tie-breaker.
+-}
+instance Ord (AssetCount GYValue) where
+  compare = comparing projection
+   where
+    projection (AssetCount v) = (valueTotalAssets v, v)
+
+newtype AssetCount a = AssetCount
+  {unAssetCount :: a}
+  deriving (Eq, Show)
+
+--------------------------------------------------------------------------------
 -- Utility functions
 --------------------------------------------------------------------------------
 
@@ -1209,3 +2035,13 @@ mapMaybe predicate (x :| xs) = go (x : xs)
     case predicate a of
       Just b -> b : go as
       Nothing -> go as
+
+naturalMonus :: Natural -> Natural -> Natural
+naturalMonus a b
+  | a >= b = a - b
+  | otherwise = 0
+
+naturalReductive :: Natural -> Natural -> Maybe Natural
+naturalReductive a b
+  | a >= b = Just (a - b)
+  | otherwise = Nothing